requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
sqlalchemy==2.0.23
pydantic==2.5.1
langchain==0.0.337
pypdf==3.17.1
python-dotenv==1.0.0
google-generativeai>=0.3.0  
faiss-cpu==1.7.4
google.ai.generativelanguage==0.6.15
google-api-core>=2.11.0
google-api-python-client
google-auth>=2.15.0
protobuf
tqdm
typing-extensions